Fix some warnings on Windows (#2560)
1. Enable warning "4503" # Decorated name length exceeded.
2. Enable warning "4146" # unary minus operator applied to unsigned type.
3. Enable float64 support for the Softmax operator
4. Enable compliance checks for Windows x86 32bits build
5. Use TryBatchParallelFor to replace some fallback code in mlas pooling.cc
6. Fix Android CI pipeline.

+++ onnxruntime/core/framework/ort_value_tensor_slicer.h
@@ -121,18 +121,18 @@ class OrtValueTensorSlicer {
     mutable OrtValue current_;
   };
 
-  Iterator begin() const noexcept { return Iterator(*ort_value_, slice_dimension_, dim0_offset_, 0); }
+  Iterator begin() const noexcept { return Iterator(*ort_value_, static_cast<size_t>(slice_dimension_), static_cast<size_t>(dim0_offset_), 0); }
   Iterator end() const noexcept {
-    return Iterator(*ort_value_, slice_dimension_, dim0_offset_, std::numeric_limits<int64_t>::max());
+    return Iterator(*ort_value_, static_cast<size_t>(slice_dimension_), static_cast<size_t>(dim0_offset_), std::numeric_limits<int64_t>::max());
   }
 
   Iterator rbegin() const noexcept {
-    return Iterator(*ort_value_, slice_dimension_, dim0_offset_, std::numeric_limits<int64_t>::max(),
+    return Iterator(*ort_value_, static_cast<size_t>(slice_dimension_), static_cast<size_t>(dim0_offset_), std::numeric_limits<int64_t>::max(),
                     Iterator::Direction::kReverse);
   }
 
   Iterator rend() const noexcept {
-    return Iterator(*ort_value_, slice_dimension_, dim0_offset_, -1, Iterator::Direction::kReverse);
+    return Iterator(*ort_value_, static_cast<size_t>(slice_dimension_), static_cast<size_t>(dim0_offset_), -1, Iterator::Direction::kReverse);
   }
 
  private:
